+++ apm-sniffer/apm-sdk-plugin/dubbo-2.7.x-plugin/src/main/resources/skywalking-plugin.def
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-dubbo=org.apache.skywalking.apm.plugin.dubbo.DubboInstrumentation
+dubbo=org.apache.skywalking.apm.plugin.asf.dubbo.DubboInstrumentation
